+++ web/package.json
@@ -14,6 +14,7 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.37",
     "@types/react-dom": "^16.9.0",
+    "@types/react-responsive": "^8.0.2",
     "@types/react-router-dom": "^5.1.5",
     "@types/styled-components": "^5.1.0",
     "@types/to-camel-case": "^1.0.0",
@@ -26,6 +27,7 @@
     "react-feather": "^2.0.8",
     "react-hook-form": "^5.7.2",
     "react-hook-inview": "^4.2.0",
+    "react-responsive": "^8.1.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "styled-components": "^5.1.1",
